Add more logs and delete some redundant code
