+++ pkg/util/namespace.go
@@ -10,20 +10,6 @@ import (
 	kubeclient "k8s.io/client-go/kubernetes"
 )
 
-// IsNamespaceExist tells if specific already exists.
-func IsNamespaceExist(client kubeclient.Interface, namespace string) (bool, error) {
-	_, err := client.CoreV1().Namespaces().Get(context.Background(), namespace, metav1.GetOptions{})
-	if err != nil {
-		if apierrors.IsNotFound(err) {
-			return false, nil
-		}
-
-		return false, err
-	}
-
-	return true, nil
-}
-
 // CreateNamespace just try to create the namespace.
 func CreateNamespace(client kubeclient.Interface, namespaceObj *corev1.Namespace) (*corev1.Namespace, error) {
 	_, err := client.CoreV1().Namespaces().Create(context.TODO(), namespaceObj, metav1.CreateOptions{})
@@ -54,14 +40,6 @@ func EnsureNamespaceExist(client kubeclient.Interface, namespace string, dryRun 
 	namespaceObj.Name = namespace
 
 	if dryRun {
-		return namespaceObj, nil
-	}
-
-	exist, err := IsNamespaceExist(client, namespace)
-	if err != nil {
-		return nil, fmt.Errorf("failed to check if namespace exist. namespace: %s, error: %v", namespace, err)
-	}
-	if exist {
 		return namespaceObj, nil
 	}
