+++ pkg/util/rbac_test.go
@@ -115,126 +115,6 @@ func TestCreateClusterRoleBinding(t *testing.T) {
 	}
 }
 
-func TestCreateOrUpdateClusterRole(t *testing.T) {
-	type args struct {
-		client      kubernetes.Interface
-		clusterRole *rbacv1.ClusterRole
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "not exist and create",
-			args: args{
-				client:      fake.NewSimpleClientset(),
-				clusterRole: makeClusterRole("test"),
-			},
-			wantErr: false,
-		},
-		{
-			name: "already exist and update",
-			args: args{
-				client:      fake.NewSimpleClientset(makeClusterRole("test")),
-				clusterRole: makeClusterRole("test"),
-			},
-			wantErr: false,
-		},
-		{
-			name: "create error",
-			args: args{
-				client: func() kubernetes.Interface {
-					c := fake.NewSimpleClientset()
-					c.PrependReactor("create", "*", errorAction)
-					return c
-				}(),
-				clusterRole: makeClusterRole("test"),
-			},
-			wantErr: true,
-		},
-		{
-			name: "update error",
-			args: args{
-				client: func() kubernetes.Interface {
-					c := fake.NewSimpleClientset(makeClusterRole("test"))
-					c.PrependReactor("update", "*", errorAction)
-					return c
-				}(),
-				clusterRole: makeClusterRole("test"),
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := CreateOrUpdateClusterRole(tt.args.client, tt.args.clusterRole); (err != nil) != tt.wantErr {
-				t.Errorf("CreateOrUpdateClusterRole() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func TestCreateOrUpdateClusterRoleBinding(t *testing.T) {
-	type args struct {
-		client             kubernetes.Interface
-		clusterRoleBinding *rbacv1.ClusterRoleBinding
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "not exist and create",
-			args: args{
-				client:             fake.NewSimpleClientset(),
-				clusterRoleBinding: makeClusterRoleBinding("test"),
-			},
-			wantErr: false,
-		},
-		{
-			name: "already exist and update",
-			args: args{
-				client:             fake.NewSimpleClientset(makeClusterRole("test")),
-				clusterRoleBinding: makeClusterRoleBinding("test"),
-			},
-			wantErr: false,
-		},
-		{
-			name: "create error",
-			args: args{
-				client: func() kubernetes.Interface {
-					c := fake.NewSimpleClientset()
-					c.PrependReactor("create", "*", errorAction)
-					return c
-				}(),
-				clusterRoleBinding: makeClusterRoleBinding("test"),
-			},
-			wantErr: true,
-		},
-		{
-			name: "update error",
-			args: args{
-				client: func() kubernetes.Interface {
-					c := fake.NewSimpleClientset(makeClusterRoleBinding("test"))
-					c.PrependReactor("update", "*", errorAction)
-					return c
-				}(),
-				clusterRoleBinding: makeClusterRoleBinding("test"),
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := CreateOrUpdateClusterRoleBinding(tt.args.client, tt.args.clusterRoleBinding); (err != nil) != tt.wantErr {
-				t.Errorf("CreateOrUpdateClusterRoleBinding() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
 func TestDeleteClusterRole(t *testing.T) {
 	type args struct {
 		client kubernetes.Interface
